@@ -6,14 +6,12 @@ authors = []
 readme = "README.md"
 
 [tool.poetry.dependencies]
-python = "^3.10"
+python = "^3.11"
 
-Flask-SQLAlchemy = "~=2.4"  
 MarkupSafe = "*"
-SQLAlchemy = "~=1.4" 
+sqlalchemy = "^2.0.26" 
 bcrypt = "*"
 cloudpathlib = {extras = ["gs"], version = "*"}
-werkzeug = "~=2.3"
 flask-rangerequest = "^0.0.2"
 flask-s3 = "*"
 flask-wtf = "*"
@@ -31,15 +29,15 @@ opentelemetry-propagator-gcp = "^1.3.0"
 opentelemetry-sdk = "^1.13.0"
 
 mysqlclient = ">=2.1"
-pydantic = "*"
+pydantic = "1.*"
 python-dateutil = "*"
 validators = "*"
 types-python-dateutil = "^2.8.19.12"
 pg8000 = "^1.30.1"
 lxml = "^4.9.2"
 xmltodict = "^0.13.0"
-arxiv-base = {git = "https://github.com/arXiv/arxiv-base.git", rev = "1a6035ab"}
-flask = "~=2.2.2"
+arxiv-base = {git = "https://github.com/arXiv/arxiv-base.git", rev = "e7b8bbac6e57850bf244f81a8cd3b6e40a48b47c"}
+flask = "^3.0.2"
 google-cloud-compute = "^1.14.1"
 
 [tool.poetry.dev-dependencies]
@@ -47,7 +45,6 @@ beautifulsoup4 = "*"
 coverage = "*"
 hypothesis = "*"
 mypy = "*"
-mypy-extensions = "*"
 pycodestyle = "*"
 pydocstyle = "*"
 pylint = "*"
@@ -57,6 +54,12 @@ pytest-easyread = "*"
 pytest-html = "*"
 pytest-mock = "*"
 types-requests = "^2.31.0.20240125"
+
+[tool.pytest.ini_options]
+filterwarnings = [
+    "ignore::DeprecationWarning",
+    "ignore:As the c extension couldn't be imported, `google-crc32c` is using a pure python implementation that is significantly slower. If possible, please configure a c build environment and compile the extension:RuntimeWarning",
+]
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
