@@ -1,10 +1,10 @@
 """Gets BibTeX citation for the paper."""
 from typing import Callable
 
-from browse.domain.identifier import IdentifierIsArchiveException, IdentifierException
+from arxiv.identifier import IdentifierIsArchiveException, IdentifierException
 from browse.formatting.cite import arxiv_bibtex
 from browse.services.documents import get_doc_service
-from browse.services.documents.base_documents import (
+from arxiv.document.exceptions import (
     AbsDeletedException, AbsException, AbsNotFoundException,
     AbsVersionNotFoundException)
 from flask import Response, make_response
