@@ -23,5 +23,4 @@ def fs_docs(config: dict, _: Any) -> DocMetadataService:
 def db_docs(config: dict, _: Any) -> DocMetadataService:
     """Factory function for DB backed abstract service."""
     from browse.services.documents.db_implementation.db_abs import DbDocMetadataService
-    from browse.services.database.models import db
-    return DbDocMetadataService(db)
+    return DbDocMetadataService()
