@@ -9,7 +9,7 @@ from wsgiref.handlers import format_date_time
 
 from flask import g, current_app
 
-from browse.domain.metadata import DocMetadata
+from arxiv.document.metadata import DocMetadata
 from browse.services import HasStatus
 
 
@@ -28,10 +28,10 @@ def fs_listing(config: dict, _: Any) -> "ListingService":
     return FsListingFilesService(config["DOCUMENT_LISTING_PATH"])
 
 def db_listing(config: dict, _: Any) -> "ListingService":
-    """Factory function for filesystem-based listing service."""
-    from browse.services.database import models
-    from .db_listing import DBListingService
-    return DBListingService(models.db)
+    """Factory function for DB backed listing service."""
+    from .db_listings import DBListingService
+    # maybe pass in the specific classes for the tables we need?
+    return DBListingService()
 
 def fake(config: Any, _: Any) -> "ListingService":
     """Factory function for fake listing service."""
