@@ -8,7 +8,8 @@ from typing import List, Optional
 from sqlalchemy import text
 from sqlalchemy.engine.base import Engine
 
-from browse.services.database.models import DBLaTeXMLDocuments
+from arxiv.db.models import Metadata
+
 
 def grep_f_count(filename: str, query: str) -> Optional[int]:
     """Like counting lines from grep -F "query" filename."""
@@ -33,7 +34,9 @@ def execute_sql_files(sql_files: List[str], engine: Engine) -> None:
                 for ln, line in enumerate([line.strip() for line in sql_file]):
                     lnum = ln
                     if not comment(line):
-                        engine.execute(text(line))
+                        with engine.connect() as conn:
+                            conn.execute(text(line))
+                            conn.commit()
         except Exception as err:
             raise Exception( f"Error at line {filename}:{lnum}") from err
     list(map(exec_sql, sql_files))
@@ -46,16 +49,24 @@ def path_of_for_test(rel_path: str) -> str:
 def foreign_key_check(engine, check_on:bool ):
     if check_on:
         if 'sqlite' in str(engine.url):
-            engine.execute(text('PRAGMA foreign_keys = ON;'))
+            with engine.connect() as conn:
+                conn.execute(text('PRAGMA foreign_keys = ON;'))
         else:
-            engine.execute(text('SET FOREIGN_KEY_CHECKS = 1;'))
+            with engine.connect() as conn:
+                conn.execute(text('SET FOREIGN_KEY_CHECKS = 1;'))
     else:
         if 'sqlite' in str(engine.url):
-            engine.execute(text('PRAGMA foreign_keys = OFF;'))
+            with engine.connect() as conn:
+                conn.execute(text('PRAGMA foreign_keys = OFF;'))
         else:
-            engine.execute(text('SET FOREIGN_KEY_CHECKS = 0;'))
+            with engine.connect() as conn:
+                conn.execute(text('SET FOREIGN_KEY_CHECKS = 0;'))
 
-def _populate_latexml_test_data (models):
+def _populate_latexml_test_data (db):
+    db.LaTeXMLBase.metadata.drop_all(bind=db.latexml_engine)
+    db.LaTeXMLBase.metadata.create_all(bind=db.latexml_engine)
+    db.session.commit()
+
     dt = datetime(2024, 1, 30, 15, 0, 0)
     dt = dt.replace(tzinfo=timezone.utc)
 
@@ -102,80 +113,85 @@ def _populate_latexml_test_data (models):
     }
 
     for doc in [doc1, doc2, doc3, doc4]:
-        models.db.session.add(DBLaTeXMLDocuments(**doc))
-    models.db.session.commit()
+        db.session.add(db.models.DBLaTeXMLDocuments(**doc))
+    db.session.commit()
 
-def populate_test_database(drop_and_create: bool, models):
+def populate_test_database(drop_and_create: bool, db):
     """Initialize the browse tables."""
     if drop_and_create:
-        foreign_key_check(models.db.engine, False)
-        models.db.drop_all()
-        models.db.session.commit()
-        foreign_key_check(models.db.engine, True)
-        models.db.create_all()
-        models.db.session.commit()
+        foreign_key_check(db.engine, False)
+        db.metadata.drop_all(bind=db.engine)
+        db.session.commit()
+        foreign_key_check(db.engine, True)
+        db.metadata.create_all(bind=db.engine)
+        db.session.commit()
 
     # Member institution data
-    models.db.session.add(
-        models.MemberInstitution(
+    db.session.add(
+        db.models.MemberInstitution(
             id=1, name='Localhost University', label='Localhost University'),
     )
-    models.db.session.add(models.MemberInstitutionIP(
+    db.session.add(db.models.MemberInstitutionIP(
         id=1, sid=1, start=2130706433, end=2130706433, exclude=0))
 
     # Intentionally add another insitution for the same loopback IP as above
-    models.db.session.add(
-        models.MemberInstitution(
+    db.session.add(
+        db.models.MemberInstitution(
             id=2, name='Loopback University', label='Loopback University'),
     )
-    models.db.session.add(models.MemberInstitutionIP(
+    db.session.add(db.models.MemberInstitutionIP(
         id=2, sid=2, start=2130706433, end=2130706433, exclude=0))
 
-    inst_cornell = models.MemberInstitution(
+    inst_cornell = db.models.MemberInstitution(
         id=3,
         name='Cornell University',
         label='Cornell University'
     )
-    models.db.session.add(inst_cornell)
+    db.session.add(inst_cornell)
 
-    inst_cornell_ip = models.MemberInstitutionIP(
+    inst_cornell_ip = db.models.MemberInstitutionIP(
         id=3,
         sid=inst_cornell.id,
         start=2152988672,  # 128.84.0.0
         end=2153054207,    # 128.84.255.255
         exclude=0
     )
-    models.db.session.add(inst_cornell_ip)
+    db.session.add(inst_cornell_ip)
 
     inst_cornell_ip_exclude = \
-        models.MemberInstitutionIP(
+        db.models.MemberInstitutionIP(
             id=4,
             sid=inst_cornell.id,
             start=2152991233,  # 128.84.10.1
             end=2152991242,    # 128.84.10.10
             exclude=1
         )
-    models.db.session.add(inst_cornell_ip_exclude)
+    db.session.add(inst_cornell_ip_exclude)
 
-    inst_other = models.MemberInstitution(
+    inst_other = db.models.MemberInstitution(
         id=5,
         name='Other University',
         label='Other University'
     )
-    models.db.session.add(inst_other)
+    db.session.add(inst_other)
 
-    inst_other_ip = models.MemberInstitutionIP(
+    inst_other_ip = db.models.MemberInstitutionIP(
         id=5,
         sid=inst_other.id,
         start=2152991236,  # 128.84.10.4
         end=2152991242,    # 128.84.10.10
         exclude=0
     )
-    models.db.session.add(inst_other_ip)
+    db.session.add(inst_other_ip)
 
-    models.db.session.commit()
     sql_files: List[str] = glob.glob('./tests/data/db/sql/*.sql')
-    foreign_key_check(models.db.engine, False)
-    execute_sql_files(sql_files, models.db.engine)
+    foreign_key_check(db.engine, False)
+    execute_sql_files(sql_files, db.engine)
+    db.session.commit()
 
-    _populate_latexml_test_data(models)
+    print ("FIRST METADATA:")
+    print (db.session.query(Metadata).first())
+
+    _populate_latexml_test_data(db)
+
+ABS_FILES = path_of_for_test('data/abs_files')
