@@ -1,7 +1,7 @@
 from datetime import date
 
-from tests.test_fs_abs_parser import ABS_FILES
-from browse.services.documents.fs_implementation.parse_abs import parse_abs_file
+from tests import ABS_FILES
+from arxiv.document.parse_abs import parse_abs_file
 from browse.formatting.external_refs_cits import include_inspire_link, \
     get_orig_publish_date
 
