@@ -2,13 +2,13 @@
 
 from unittest import TestCase
 
-from arxiv.util.authors import split_authors
+from arxiv.authors import split_authors
 
-from browse.services.documents.fs_implementation.parse_abs import parse_abs_file
+from arxiv.document.parse_abs import parse_abs_file
 from browse.formatting.search_authors import queries_for_authors, split_long_author_list
 from tests import path_of_for_test
 
-from browse.domain import metadata
+from arxiv.document.metadata import DocMetadata
 
 
 class TestAuthorLinkCreation(TestCase):
@@ -51,7 +51,7 @@ class TestAuthorLinkCreation(TestCase):
     def test_split_long_author_list(self):
         f1 = path_of_for_test(
             'data/abs_files/ftp/arxiv/papers/1411/1411.4413.abs')
-        meta: metadata = parse_abs_file(filename=f1)
+        meta = parse_abs_file(filename=f1)
         alst = split_long_author_list(
             queries_for_authors(str(meta.authors)), 20)
         self.assertIs(type(alst), tuple)
@@ -64,7 +64,7 @@ class TestAuthorLinkCreation(TestCase):
     def test_split_with_collaboration(self):
         f1 = path_of_for_test(
             'data/abs_files/ftp/arxiv/papers/0808/0808.4142.abs')
-        meta: metadata = parse_abs_file(filename=f1)
+        meta = parse_abs_file(filename=f1)
 
         split = split_authors(str(meta.authors))
         self.assertListEqual(
@@ -78,7 +78,7 @@ class TestAuthorLinkCreation(TestCase):
         """Test odd author list that shows '0 additional authors' ARXIVNG-2083"""
         f1 = path_of_for_test(
             'data/abs_files/ftp/arxiv/papers/1902/1902.05884.abs')
-        meta: metadata = parse_abs_file(filename=f1)
+        meta = parse_abs_file(filename=f1)
         alst = split_long_author_list(
             queries_for_authors(str(meta.authors)), 100)
 
